requirements.txt
discord.py 
youtube-dl==2020.12.2 
pafy 
PyNaCl 
ffmpeg-python 
asyncio